agent/requirements.txt
griptape[drivers-prompt-google]==1.3.5
python-dotenv